+++ kpf/utils/SetObserverFromSchedule.py
@@ -1,6 +1,4 @@
 import time
-import requests
-import json
 from datetime import datetime, timedelta
 
 import ktl
@@ -8,32 +6,9 @@ import ktl
 from kpf.KPFTranslatorFunction import KPFTranslatorFunction
 from kpf import (log, KPFException, FailedPreCondition, FailedPostCondition,
                  FailedToReachDestination, check_input)
+from kpf.utils.telsched import get_schedule
 from kpf.spectrograph.SetObserver import SetObserver
 from kpf.spectrograph.SetProgram import SetProgram
-
-
-##-----------------------------------------------------------------------------
-## Functions to interact with telescope DB
-##-----------------------------------------------------------------------------
-def querydb(req):
-    '''A simple wrapper to form a generic API level query to the telescope
-    schedule web API.  Returns a JSON object with the result of the query.
-    '''
-    url = f"https://www.keck.hawaii.edu/software/db_api/telSchedule.php?{req}"
-    r = requests.get(url)
-    return json.loads(r.text)
-
-
-def get_schedule(date, tel):
-    '''Use the querydb function and getSchedule of the telescope schedule web
-    API with arguments for date and telescope number.  Returns a JSON object
-    with the schedule result.
-    '''
-    if tel not in [1,2]:
-        raise KPFError(f"Telescope number in query must be 1 or 2")
-    req = f"cmd=getSchedule&date={date}&telnr={tel}"
-    result = querydb(req)
-    return result
 
 
 ##-----------------------------------------------------------------------------
